requirements.txt
pytorch>=1.7.0
torch-cluster>=1.5.8
torch-geometric>=1.6.3
torch-scatter>-=2.0.5
torch-sparse>=0.6.8
scipy>=1.5.4